controller produces .csv file

+++ src/main/java/com/verygoodbank/tes/web/model/TradeRequest.java
@@ -1,7 +1,7 @@
 package com.verygoodbank.tes.web.model;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
-public record TradeRequest(String productId, BigInteger price, String date, String currency) {
+public record TradeRequest(String productId, BigDecimal price, String date, String currency) {
 
 }
